+++ src/utils/kiss_fft.c
@@ -306,7 +306,7 @@ static void kf_bfly_generic(
             int twidx=0;
             Fout[ k ] = scratchbuf[0];
             for (q=1;q<p;++q ) {
-                twidx += fstride * k;
+                twidx += (int)(fstride * k);
                 if (twidx>=Norig) twidx-=Norig;
                 C_MUL(t,scratchbuf[q] , twiddles[twidx] );
                 C_ADDTO( Fout[ k ] ,t);
@@ -411,7 +411,7 @@ void kf_work(
           }
        }*/
     }else{
-       kf_work( Fout , f, fstride*p, in_stride, factors,st, N*p, fstride*in_stride, m);
+       kf_work( Fout , f, fstride*p, in_stride, factors,st, N*p, (int)(fstride*in_stride), m);
     }
